splitting up transaction flow components
